# docs/swagger.yaml
definitions:
...
  requests.CreateMerchantPolicyRequest:
    properties:
      description:
        type: string
      merchant_id:
        type: integer
      policy_type:
        type: string
      title:
        type: string
    required:
    - description
    - merchant_id
    - policy_type
    - title
    type: object
...
  response.ApiResponseMerchant:
    properties:
      data:
        $ref: '#/definitions/response.MerchantResponse'
      message:
        type: string
      status:
        type: string
    type: object
  response.ApiResponseMerchantAll:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
...
  response.ApiResponseMerchantDelete:
    properties:
      message:
        type: string
      status:
        type: string
    type: object
  response.ApiResponseMerchantDeleteAt:
    properties:
      data:
        $ref: '#/definitions/response.MerchantResponseDeleteAt'
      message:
        type: string
      status:
        type: string
    type: object
...
  response.ApiResponseMerchantPolicies:
    properties:
      data:
        $ref: '#/definitions/response.MerchantPoliciesResponse'
      message:
        type: string
      status:
        type: string
    type: object
  response.ApiResponseMerchantPoliciesDeleteAt:
    properties:
      data:
        $ref: '#/definitions/response.MerchantPoliciesResponseDeleteAt'
      message:
        type: string
      status:
        type: string
    type: object
...
  response.ApiResponsePaginationMerchant:
    properties:
      data:
        items:
          $ref: '#/definitions/response.MerchantResponse'
        type: array
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.PaginationMeta'
      status:
        type: string
    type: object
...
  response.ApiResponsePaginationMerchantPolicies:
    properties:
      data:
        items:
          $ref: '#/definitions/response.MerchantPoliciesResponse'
        type: array
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.PaginationMeta'
      status:
        type: string
    type: object
  response.ApiResponsePaginationMerchantPoliciesDeleteAt:
    properties:
      data:
        items:
          $ref: '#/definitions/response.MerchantPoliciesResponseDeleteAt'
        type: array
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.PaginationMeta'
      status:
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      status:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/merchant:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all merchant
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of merchant
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationMerchant'
        "500":
          description: Failed to retrieve merchant data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find all merchant
      tags:
      - Merchant
...
  /api/merchant-policy:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all merchant
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of merchant
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationMerchantPolicies'
        "500":
          description: Failed to retrieve merchant data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find all merchant
      tags:
      - MerchantPolicy
  /api/merchant-policy/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a merchant by ID
      parameters:
      - description: merchant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: merchant data
          schema:
            $ref: '#/definitions/response.ApiResponseMerchant'
        "400":
          description: Invalid merchant ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to retrieve merchant data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find merchant by ID
      tags:
      - MerchantPolicy
  /api/merchant-policy/active:
    get:
      consumes:
      - application/json
      description: Retrieve a list of active merchant
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of active merchant
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationMerchantPoliciesDeleteAt'
        "500":
          description: Failed to retrieve merchant data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Retrieve active merchant
      tags:
      - MerchantPolicy
  /api/merchant-policy/create:
    post:
      consumes:
      - application/json
      description: Create a new merchant policy with the provided details
      parameters:
      - description: Create merchant policy request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.CreateMerchantPolicyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created merchant policy
          schema:
            $ref: '#/definitions/response.ApiResponseMerchantPolicies'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to create merchant policy
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a new merchant policy
      tags:
      - MerchantPolicy
  /api/merchant-policy/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Permanently delete a merchant record by its ID.
      parameters:
      - description: merchant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted merchant record permanently
          schema:
            $ref: '#/definitions/response.ApiResponseMerchantDelete'
        "400":
          description: 'Bad Request: Invalid ID'
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 'Failed to delete merchant:'
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Permanently delete a merchant
      tags:
      - MerchantPolicy
  /api/merchant-policy/delete/all:
    post:
      consumes:
      - application/json
      description: Permanently delete a merchant record by its ID.
      parameters:
      - description: merchant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted merchant record permanently
          schema:
            $ref: '#/definitions/response.ApiResponseMerchantAll'
        "400":
          description: 'Bad Request: Invalid ID'
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: 'Failed to delete merchant:'
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Permanently delete a merchant
      tags:
      - MerchantPolicy
  /api/merchant-policy/restore/{id}:
    post:
      consumes:
      - application/json
      description: Restore a trashed merchant record by its ID.
      parameters:
      - description: Merchant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully restored merchant
          schema:
            $ref: '#/definitions/response.ApiResponseMerchantPoliciesDeleteAt'
        "400":
          description: Invalid merchant ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to restore merchant
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Restore a trashed merchant
      tags:
      - MerchantPolicy
  /api/merchant-policy/restore/all:
    post:
      consumes:
      - application/json
      description: Restore a trashed merchant record by its ID.
      parameters:
      - description: merchant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully restored merchant all
          schema:
            $ref: '#/definitions/response.ApiResponseMerchantAll'
        "400":
          description: Invalid merchant ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to restore merchant
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Restore a trashed merchant
      tags:
      - MerchantPolicy
  /api/merchant-policy/trashed:
    get:
      consumes:
      - application/json
      description: Retrieve a list of trashed merchant records
      produces:
      - application/json
      responses:
        "200":
          description: List of trashed merchant data
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationMerchantPoliciesDeleteAt'
        "500":
          description: Failed to retrieve merchant data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Retrieve trashed merchant
      tags:
      - MerchantPolicy
  /api/merchant-policy/trashed/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a trashed merchant record by its ID.
      parameters:
      - description: Merchant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved trashed merchant
          schema:
            $ref: '#/definitions/response.ApiResponseMerchantPoliciesDeleteAt'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to retrieve trashed merchant
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Retrieve a trashed merchant
      tags:
      - MerchantPolicy
  /api/merchant-policy/update/{id}:
    post:
      consumes:
      - application/json
      description: Update an existing merchant policy with the provided details
      parameters:
      - description: Merchant Policy ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update merchant policy request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.UpdateMerchantPolicyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated merchant policy
          schema:
            $ref: '#/definitions/response.ApiResponseMerchantPolicies'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to update merchant policy
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Update an existing merchant policy
      tags:
      - MerchantPolicy
  /api/merchant/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a merchant by ID
      parameters:
      - description: merchant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: merchant data
          schema:
            $ref: '#/definitions/response.ApiResponseMerchant'
        "400":
          description: Invalid merchant ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to retrieve merchant data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Find merchant by ID
      tags:
      - Merchant
  /api/merchant/active:
    get:
      consumes:
      - application/json
      description: Retrieve a list of active merchant
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        name: page_size
        type: integer
      - description: Search query
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of active merchant
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationMerchantDeleteAt'
        "500":
          description: Failed to retrieve merchant data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Retrieve active merchant
      tags:
      - Merchant
  /api/merchant/create:
    post:
      consumes:
      - application/json
      description: Create a new merchant with the provided details
      parameters:
      - description: Create merchant request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/requests.CreateMerchantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created merchant
          schema:
            $ref: '#/definitions/response.ApiResponseMerchant'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to create merchant
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Create a new merchant
      tags:
      - Merchant
...
  /api/merchant/restore/{id}:
    post:
      consumes:
      - application/json
      description: Restore a trashed merchant record by its ID.
      parameters:
      - description: Merchant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully restored merchant
          schema:
            $ref: '#/definitions/response.ApiResponseMerchantDeleteAt'
        "400":
          description: Invalid merchant ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to restore merchant
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Restore a trashed merchant
      tags:
      - Merchant
  /api/merchant/restore/all:
    post:
      consumes:
      - application/json
      description: Restore a trashed merchant record by its ID.
      parameters:
      - description: merchant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully restored merchant all
          schema:
            $ref: '#/definitions/response.ApiResponseMerchantAll'
        "400":
          description: Invalid merchant ID
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to restore merchant
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Restore a trashed merchant
      tags:
      - Merchant
  /api/merchant/trashed:
    get:
      consumes:
      - application/json
      description: Retrieve a list of trashed merchant records
      produces:
      - application/json
      responses:
        "200":
          description: List of trashed merchant data
          schema:
            $ref: '#/definitions/response.ApiResponsePaginationMerchantDeleteAt'
        "500":
          description: Failed to retrieve merchant data
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Retrieve trashed merchant
      tags:
      - Merchant
  /api/merchant/trashed/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a trashed merchant record by its ID.
      parameters:
      - description: Merchant ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved trashed merchant
          schema:
            $ref: '#/definitions/response.ApiResponseMerchantDeleteAt'
        "400":
          description: Invalid request body or validation error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Failed to retrieve trashed merchant
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - Bearer: []
      summary: Retrieve a trashed merchant
      tags:
      - Merchant
...
swagger: "2.0"